BC-521: Move result into upload controller
