+++ src/main/resources/templates/bulk-upload.html
@@ -21,7 +21,7 @@
                             <span class="govuk-details__summary-text" th:text="#{bulkUpload.csvFormatSummary}"></span>
                         </summary>
                         <div class="govuk-details__text">
-                            TODO
+                            TODO: To be populated by BC-483
                         </div>
                     </details>
